requirements.txt
ebooklib>=0.17.1
Pillow>=10.0.0
requests>=2.31.0
click>=8.1.3
tqdm>=4.66.1
beautifulsoup4>=4.10.0
aiohttp>=3.8.1
colorama>=0.4.6
